# ENH: use algorithm A5 of philippe for sampling from left-truncated gamma