closes #18