@zoj613 2021-03-11
Im not sure this is worth merging. There is no significant change in speed compared to the main branch.

Main branch:
```python
In [1]: from polyagamma import polyagamma

In [2]: import numpy as np

In [9]: rng = np.random.default_rng(1)

In [10]: %timeit polyagamma(20, 5, method="devroye", random_state=rng, size=20000)
59.1 ms ± 84.5 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)

In [11]: rng = np.random.default_rng(1)

In [12]: %timeit polyagamma(20, 5, method="alternate", random_state=rng, size=20000)
25.3 ms ± 541 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)

In [15]: rng = np.random.default_rng(1)

In [16]: %timeit polyagamma(20, 5, method="saddle", random_state=rng, size=20000)
33.9 ms ± 353 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
```


vs this PR:
```python
In [1]: from polyagamma import polyagamma

In [2]: import numpy as np

In [3]: rng = np.random.default_rng(1)

In [4]: %timeit polyagamma(20, 5, method="saddle", random_state=rng, size=20000)
33.4 ms ± 83.9 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)

In [5]: rng = np.random.default_rng(1)

In [6]: %timeit polyagamma(20, 5, method="alternate", random_state=rng, size=20000)
25 ms ± 846 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)

In [7]: rng = np.random.default_rng(1)

In [8]: %timeit polyagamma(20, 5, method="devroye", random_state=rng, size=20000)
59.2 ms ± 644 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
```

I think this is mainly due to the fact that algorithm A5 requires calculation of the mixture weights on every call to the truncated gamma sampler. The first weight uses special functions which are a little expensive. So the nature of the algorithms there is no obvious way to take advantage of the high acceptance rate A5 to improve speed in a significant way. So maybe this PR is not worthwhile.

I'll likely leave this open for reference purposes.