@zoj613 2021-03-11
When `z` is small, this PR is a tiny bit faster

Main branch:
```python
In [4]: rng = np.random.default_rng(1)

In [5]: %timeit polyagamma(20, 0.5, method="alternate", random_state=rng, size=20000)
129 ms ± 658 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)

In [6]: rng = np.random.default_rng(1)

In [7]: %timeit polyagamma(20, 0.5, method="saddle", random_state=rng, size=20000)
41.3 ms ± 555 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)

In [8]: rng = np.random.default_rng(1)

In [9]: %timeit polyagamma(20, 0.5, method="devroye", random_state=rng, size=20000)
50.2 ms ± 80.9 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
```
vs

This PR:
```python
In [3]: rng = np.random.default_rng(1)

In [4]: %timeit polyagamma(20, 0.5, method="devroye", random_state=rng, size=20000)
50 ms ± 244 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)

In [8]: rng = np.random.default_rng(1)

In [9]: %timeit polyagamma(20, 0.5, method="saddle", random_state=rng, size=20000)
41.2 ms ± 439 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)

In [10]: rng = np.random.default_rng(1)

In [11]: %timeit polyagamma(20, 0.5, method="alternate", random_state=rng, size=20000)
128 ms ± 89.9 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
```